Re-stack on updated run_as base and fix SDK CI
- Rebase onto squashed cursor/workflow-run-as-execution-authority-d43b
- Fix TypeScript type errors in workflow-authoring and workflow-public exports
- Merge authoring golden tests into existing workflow builder test suites
- Fix Python ruff unused imports and Rust formatting
