+++ sdk/rust/src/lib.rs
@@ -141,16 +141,16 @@ pub use secrets_provider::SecretsProvider;
 pub use tonic::codegen::async_trait;
 pub use workflow::Workflow;
 pub use workflow_authoring::{
-    apply_workflow_definition, build_workflow_from_lowering_case, canonical_workflow_definition_spec,
-    define_workflow, event, load_workflow_lowering_contract, resolve_workflow_definition_spec,
-    resolve_workflow_definition_spec_from_builder, schedule, workflow_ref_array, workflow_ref_input,
-    workflow_ref_literal, workflow_ref_object, workflow_ref_signal, workflow_ref_step_input,
-    workflow_ref_step_output, workflow_ref_template, DefineWorkflowOptions, WorkflowActivationConfig,
-    WorkflowActivationScope, WorkflowBuilder, WorkflowDefinitionSpecOrBuilder,
-    WorkflowEventActivationConfig, WorkflowEventActivationOptions, WorkflowEventScope,
-    WorkflowLoweringCase, WorkflowScheduleActivationConfig, WorkflowScheduleActivationOptions,
-    WorkflowStepAgentConfig, WorkflowStepAgentMessageConfig, WorkflowStepAppConfig,
-    WorkflowStepConfig, WorkflowStepScope, WorkflowStepWhenConfig,
+    DefineWorkflowOptions, WorkflowActivationConfig, WorkflowActivationScope, WorkflowBuilder,
+    WorkflowDefinitionSpecOrBuilder, WorkflowEventActivationConfig, WorkflowEventActivationOptions,
+    WorkflowEventScope, WorkflowLoweringCase, WorkflowScheduleActivationConfig,
+    WorkflowScheduleActivationOptions, WorkflowStepAgentConfig, WorkflowStepAgentMessageConfig,
+    WorkflowStepAppConfig, WorkflowStepConfig, WorkflowStepScope, WorkflowStepWhenConfig,
+    apply_workflow_definition, build_workflow_from_lowering_case,
+    canonical_workflow_definition_spec, define_workflow, event, load_workflow_lowering_contract,
+    resolve_workflow_definition_spec, resolve_workflow_definition_spec_from_builder, schedule,
+    workflow_ref_array, workflow_ref_input, workflow_ref_literal, workflow_ref_object,
+    workflow_ref_signal, workflow_ref_step_input, workflow_ref_step_output, workflow_ref_template,
 };
 pub use workflow_provider::{
     ApplyWorkflowProviderDefinitionRequest, BoundWorkflowTarget, CancelWorkflowProviderRunRequest,
