+++ sdk/typescript/src/providers/workflow-public.ts
@@ -11,26 +11,17 @@ export {
   WorkflowRunStatus,
   WorkflowStepStatus,
   WorkflowValueError,
-  WorkflowBuilder,
   boundWorkflowTarget,
   boundWorkflowTargetFromTarget,
   boundWorkflowTargetInputFromTarget,
-  applyWorkflowDefinition,
-  buildWorkflowFromLoweringCase,
-  canonicalWorkflowDefinitionSpec,
   createWorkflowProviderService,
-  defineWorkflow,
   defineWorkflowProvider,
   evaluateWorkflowStepInputs,
   evaluateWorkflowValue,
-  event,
   isWorkflowProvider,
-  lowerWorkflowValueNode,
   parseWorkflowRunContext,
   pathValue,
   renderWorkflowTemplate,
-  resolveWorkflowDefinitionSpec,
-  schedule,
   workflowActivation,
   workflowAgentMessage,
   workflowDefinition,
@@ -84,20 +75,6 @@ export {
   type WorkflowActivation,
   type WorkflowActivationTriggerKind,
   type WorkflowAgentMessage,
-  type ActivationConfig,
-  type ActivationScope,
-  type DefineWorkflowOptions,
-  type EventActivationConfig,
-  type EventActivationOptions,
-  type EventScope,
-  type ScheduleActivationConfig,
-  type ScheduleActivationOptions,
-  type StepAgentConfig,
-  type StepAgentMessageConfig,
-  type StepAppConfig,
-  type StepConfig,
-  type StepScope,
-  type StepWhenConfig,
   type WorkflowDefinition,
   type WorkflowDefinitionSpec,
   type WorkflowEvalContext,
@@ -131,3 +108,28 @@ export {
   type WorkflowValue,
   type WorkflowValueKind,
 } from "./workflow.ts";
+export {
+  WorkflowBuilder,
+  applyWorkflowDefinition,
+  buildWorkflowFromLoweringCase,
+  canonicalWorkflowDefinitionSpec,
+  defineWorkflow,
+  event,
+  lowerWorkflowValueNode,
+  resolveWorkflowDefinitionSpec,
+  schedule,
+  type ActivationConfig,
+  type ActivationScope,
+  type DefineWorkflowOptions,
+  type EventActivationConfig,
+  type EventActivationOptions,
+  type EventScope,
+  type ScheduleActivationConfig,
+  type ScheduleActivationOptions,
+  type StepAgentConfig,
+  type StepAgentMessageConfig,
+  type StepAppConfig,
+  type StepConfig,
+  type StepScope,
+  type StepWhenConfig,
+} from "../workflow-authoring.ts";
